==> 02_Creating_Seasonal_Parameters.Rmd <==
---
title: "Creating Valley fever Seasonal Timing Parameters"
author: "Simon Camponuri"
date: "1/20/2022"
output: html_document
editor_options: 
  chunk_output_type: console
---

Load libraries
```{r}
library(stringr)
library(stringi)
library(forcats)
library(tidyverse)
library(segmented)
library(tidyverse)
library(sf)
library(cowplot)
library(msm)
setwd('~/Desktop/Season Timing')
```

Necessary functions
```{r}
# estimate epidemic start (or end) time for an increase (decrease) 
# half epidemic curve using broken-line model
#' @param half the half epidemic curve, expressed as time series of observations
segment.half.ec <- function(half) {
  require(segmented)
  fit.lm <- lm(y ~ t, data = half)
  fit.seg <- segmented(fit.lm, seg.Z = ~t,
                       npsi = 1)
  idx1 <- which.min(abs(half$t - floor(fit.seg$psi[2])))
  idx2 <- which.min(abs(half$t - ceiling(fit.seg$psi[2])))
  # linearly interpolate to obtain the y value at epi.start
  require(Hmisc)
  res <- approxExtrap(x = half$t[c(idx1, idx2)], y = half$y[c(idx1, idx2)], 
                      xout = fit.seg$psi[2])
  yval <- res$y
  return(c(fit.seg$psi, yval))
}

# estimate epidemic parameters for a whole epidemic curve using broken-line model
#' @param ec the whole epidemic curve, expressed as time series of observations
segment.ec <- function(ec) {
  pk.idx <- which.max(ec$y)
  epi.peak <- ec$t[pk.idx]
  epi.peak.num <- ec$y[pk.idx]
  half1 <- ec[1:pk.idx, ]
  half2 <- ec[pk.idx:nrow(ec), ]
  psi <- segment.half.ec(half1)
  epi.start <- psi[2]
  epi.start.sd <- psi[3]
  epi.start.num <- psi[4]
  psi <- segment.half.ec(half2)
  epi.end <- psi[2]
  epi.end.sd <- psi[3]
  epi.end.num <- psi[4]
  epi.duration <- epi.end - epi.start + 1
  dur.idx <- ec$t >= epi.start & ec$t <= epi.end
  # mean intensity during epidemic duration
  epi.duration.num <- sum(ec$y[dur.idx]) / sum(dur.idx)
  return(data.frame(epi.start, epi.start.num, 
                    epi.end, epi.end.num, 
                    epi.peak, epi.peak.num, 
                    epi.duration, epi.duration.num, 
                    epi.start.sd, epi.end.sd))
}

```

SRM and MCM Detection Function
```{r}
source("Code/curvature.R")
source("Code/curvature_np.R")
```


Load data
```{r}
ct.data <- read_rds('Data/timing_data.rds')
county.data <- ct.data %>% group_by(wkyr, first_day, continuous_week_onset, tyear, county1, week, tweek, month) %>%
  summarise(N = sum(N), pop = sum(CTpop)) %>% arrange(county1, first_day) %>% mutate(IR = N/pop*100000, c = paste(county1,'-', tyear))
```

Testing smoothing parameters
```{r}
# ct.data.seasonal <- ct.data %>% filter(seasonal == 1)
# #SRM
# smoothed_all_segs <- data.frame()
# for (j in 1:15) {
#   print(j)
#   all_segs <- data.frame()
# for (i in unique(ct.data.seasonal$c)) {
#   tryCatch({
#   season <- ct.data.seasonal %>% dplyr::filter(c == i)
#   smooth <- ksmooth(time(season$tweek),season$N,'normal',bandwidth=j)
#   ec <- data.frame(t = smooth$x, y = smooth$y)
#   seg <- data.frame(segment.ec(ec))
#   seg$c <- i
#   all_segs <- rbind(all_segs, seg)
# }, error=function(e){cat(i, conditionMessage(e), "\n")})
# }
# all_segs$bandwidth <- j
# smoothed_all_segs <- rbind(smoothed_all_segs, all_segs)
# }
# 
# smoothed_all_segs <- smoothed_all_segs %>% dplyr::mutate(GEOID = str_sub(c, 1, 11),
#                                              tyear = as.numeric(str_sub(c, 15,18)))
# d1 <- smoothed_all_segs %>%
#   arrange(GEOID, tyear,c) %>%
#   group_by(GEOID, tyear) %>%
#   dplyr::mutate(start_diff = abs(epi.start - lag(epi.start)),
#                 end_diff = abs(epi.end - lag(epi.end))) #%>%
# p1 <- d1 %>% group_by(bandwidth) %>% dplyr::mutate(mean_start_diff = mean(start_diff, na.rm = T)) %>%
#   ggplot(aes(as.factor(bandwidth))) + 
#   geom_point(aes(y = mean_start_diff)) + 
#  # geom_vline(xintercept = 7, linetype = 'dashed') + 
#   geom_hline(yintercept = 1, linetype = 'dashed') + 
#   labs(title = 'Segmented Regression Method',x = 'Bandwidth',y= 'Onset Date Change')+ 
#   theme_bw()
# 
# d2 <- smoothed_all_segs %>%
#   arrange(GEOID, tyear,c) %>%
#   group_by(GEOID, tyear) %>%
#   dplyr::mutate(start_diff = abs(epi.start - lag(epi.start)),
#                 end_diff = abs(epi.end - lag(epi.end)))
# 
# p2 <- d2 %>% group_by(bandwidth) %>% dplyr::mutate(mean_end_diff = mean(end_diff, na.rm = T)) %>%
#   ggplot(aes(as.factor(bandwidth))) + 
#   geom_point(aes(y = mean_end_diff)) + 
# #  geom_vline(xintercept = 7, linetype = 'dashed') + 
#     geom_hline(yintercept = 1, linetype = 'dashed') + 
#   labs(title = 'Segmented Regression Method',x = 'Bandwidth',y= 'End Date Change') + 
#   theme_bw()
# 
# library(cowplot)
# png(filename = 'bandwidth_comparison_srm.png', width = 12, height = 6, units = 'in', res = 400)
# plot_grid(p1,p2, ncol = 2)
# dev.off()
# 
# 
# #MCM
# 
# params_mcm <- data.frame()
# for (j in 1:15) {
#   all_segs <- data.frame()
# for (i in unique(ct.data.seasonal$c)) {
#   tryCatch({
#   season <- ct.data.seasonal %>% dplyr::filter(c == i)
#   smooth <- ksmooth(time(season$tweek),season$N,'normal',bandwidth=j)
#   ec <- data.frame(t = smooth$x, y = smooth$y)
#   mcm <- data.frame(curvature.ec(ec, smoothing = F))
#   mcm$c <- i
#   all_segs <- rbind(all_segs, mcm)
# }, error=function(e){cat(i, conditionMessage(e), "\n")})
# }
# all_segs$bandwidth <- j
# params_mcm <- rbind(params_mcm, all_segs)
# }
# 
# params_mcm <- params_mcm %>% dplyr::mutate(GEOID = str_sub(c, 1, 11),
#                                              tyear = as.numeric(str_sub(c, 15,18)))
# d1 <- params_mcm %>%
#   arrange(GEOID, tyear,c) %>%
#   group_by(GEOID, tyear) %>%
#   dplyr::mutate(start_diff = abs(epi.start - lag(epi.start)),
#                 end_diff = abs(epi.end - lag(epi.end))) #%>%
# p3 <- d1 %>% group_by(bandwidth) %>% dplyr::mutate(mean_start_diff = mean(start_diff, na.rm = T)) %>%
#   ggplot(aes(as.factor(bandwidth))) + geom_point(aes(y = mean_start_diff)) + 
#     geom_hline(yintercept = 1, linetype = 'dashed') + 
#   labs(title = 'Maximum Curvature Method',x = 'Bandwidth',y= 'Onset Date Change')+ theme_bw()
# 
# d2 <- params_mcm %>%
#   arrange(GEOID, tyear,c) %>%
#   group_by(GEOID, tyear) %>%
#   dplyr::mutate(start_diff = abs(epi.start - lag(epi.start)),
#                 end_diff = abs(epi.end - lag(epi.end)))
# 
# p4 <- d2 %>% group_by(bandwidth) %>% dplyr::mutate(mean_end_diff = mean(end_diff, na.rm = T)) %>%
#   ggplot(aes(as.factor(bandwidth))) + geom_point(aes(y = mean_end_diff)) + 
#     geom_hline(yintercept = 1, linetype = 'dashed') + 
#   labs(title = 'Maximum Curvature Method',x = 'Bandwidth',y= 'End Date Change') + theme_bw()
# 
# png(filename = 'bandwidth_comparison.png', width = 12, height = 10, units = 'in', res = 400)
# plot_grid(p1,p2,p3,p4, ncol = 2, nrow =2)
# dev.off()
```

Smoothing cases
```{r}
ct.data <- ct.data %>% 
  group_by(GEOID) %>% arrange(GEOID, first_day) %>% 
  mutate(smooth = ksmooth(continuous_week_onset,IR,'normal',bandwidth=8,n.points = length(continuous_week_onset))$y)

county.data <- county.data %>% 
  group_by(county1) %>% arrange(county1, first_day) %>% 
  mutate(smooth = ksmooth(continuous_week_onset,IR,'normal',bandwidth=8, n.points = length(continuous_week_onset))$y)
```

Season inclusion criteria
```{r}
model_comp <- function(x,y) {
  har_test <- lm(y ~ sin(2*pi*x/52) + cos(2*pi*x/52) + x)
  lm_test <- lm(y ~ x)
  comp <- AIC(lm_test) - AIC(har_test)
  diff <- data.frame(diff = comp, 
                     a = as.numeric(har_test$coefficients[2]),
                     b = as.numeric(har_test$coefficients[3])) %>%
    dplyr::mutate(ab = ifelse(a < 0 & b > 0, 1, 0))
  return(diff)
}


  all_diffs <- data.frame()
  for (i in unique(county.data$c)) {
    df <- county.data %>% dplyr::filter(c == i)
    if (sum(df$N) > 0) {
    diff <- model_comp(df$tweek, df$IR) 
    diff$c <- i
    all_diffs <- rbind(all_diffs, diff)
  } else {}
  }
  county_seasons <- all_diffs %>% dplyr::filter(diff > 0, ab != 1)

    all_diffs <- data.frame()
  for (i in unique(ct.data$c)) {
    df <- ct.data %>% dplyr::filter(c == i)
    if (sum(df$N) > 0) {
    diff <- model_comp(df$tweek, df$IR) 
    diff$c <- i
    all_diffs <- rbind(all_diffs, diff)
    } else {}
  }
    nrow(all_diffs[all_diffs$diff >= 2,])
    nrow(all_diffs)
    nrow(all_diffs)
    length(unique(ct.data$c))
  census_seasons <- all_diffs %>% dplyr::filter(diff > 0, ab != 1)

county.data <- county.data %>% mutate(seasonal = ifelse(c %in% county_seasons$c,1,0))
ct.data <- ct.data %>% mutate(seasonal = ifelse(c %in% census_seasons$c,1,0))
print(c(nrow(census_seasons),nrow(county_seasons)))
ct.data %>% ungroup() %>% dplyr::select(GEOID, tyear, seasonal) %>% distinct() %>% group_by(GEOID) %>% dplyr::summarize(num_years = n(), seasonal_years = sum(seasonal)) %>% mutate(any_season = ifelse(seasonal_years > 0, 1, 0)) %>% pull(any_season) %>% sum()
county.data %>% ungroup() %>% dplyr::select(county1, tyear, seasonal) %>% distinct() %>% group_by(county1) %>% dplyr::summarize(num_years = n(), seasonal_years = sum(seasonal)) 
```


Running srm onset/end detection methods
```{r}
county.srm <- data.frame()
for (i in unique(county.data$c)) {
  season <- county.data %>% dplyr::filter(c == i)
  if (any(season$seasonal == 1)) {
  tryCatch({
  ec <- data.frame(t = season$tweek, y = season$smooth)
  seg <- data.frame(segment.ec(ec))
  seg$c <- i
  county.srm <- rbind(county.srm, seg)
}, error=function(e){cat(i, conditionMessage(e), "\n")})
  } else{}
}
county.srm <- county.srm %>% mutate(county1 = str_split(c, ' - ', simplify = T)[,1],
                                    tyear = as.numeric(str_split(c, ' - ', simplify = T)[,2]))
names(county.srm)[1:10] <- paste0(names(county.srm)[1:10],'.srm')

ct.data.seasonal <- ct.data %>% dplyr::filter(seasonal == 1)
census.srm <- data.frame()
for (i in unique(ct.data.seasonal$c)) {
  season <- ct.data.seasonal %>% dplyr::filter(c == i)
  if (any(season$seasonal == 1)) {
  tryCatch({
  ec <- data.frame(t = season$tweek, y = season$smooth)
  seg <- data.frame(segment.ec(ec))
  seg$c <- i
  census.srm <- rbind(census.srm, seg)
}, error=function(e){cat(i, conditionMessage(e), "\n")})
  } else{}
}
census.srm <- census.srm %>% mutate(GEOID = str_split(c, ' - ', simplify = T)[,1],
                                    tyear = as.numeric(str_split(c, ' - ', simplify = T)[,2]))
names(census.srm)[1:10] <- paste0(names(census.srm)[1:10],'.srm')

county.mcm <- data.frame()
for (i in unique(county.data$c)) {
  season <- county.data %>% dplyr::filter(c == i)
  if (any(season$seasonal == 1)) {
  tryCatch({
  ec <- data.frame(t = season$tweek, y = season$smooth)
  mcm <- data.frame(curvature.ec(ec, n = 5, h = 5, smoothing = F))
  mcm$c <- i
  county.mcm <- rbind(county.mcm, mcm)
}, error=function(e){cat(i, conditionMessage(e), "\n")})
  } else {}
}
county.mcm <- county.mcm %>% mutate(county1 = str_split(c, ' - ', simplify = T)[,1],
                                    tyear = as.numeric(str_split(c, ' - ', simplify = T)[,2]))
names(county.mcm)[1:8] <- paste0(names(county.mcm)[1:8],'.mcm')

census.mcm <- data.frame()
for (i in unique(ct.data.seasonal$c)) {
  season <- ct.data %>% dplyr::filter(c == i)
  if (any(season$seasonal == 1)) {
  tryCatch({
  ec <- data.frame(t = season$tweek, y = season$smooth)
  mcm <- data.frame(curvature.ec(ec, n = 5, h = 5, smoothing = F))
  mcm$c <- i
  census.mcm <- rbind(census.mcm, mcm)
}, error=function(e){cat(i, conditionMessage(e), "\n")})
  } else {}
}
census.mcm <- census.mcm %>% mutate(GEOID = str_split(c, ' - ', simplify = T)[,1],
                                    tyear = as.numeric(str_split(c, ' - ', simplify = T)[,2]))
names(census.mcm)[1:8] <- paste0(names(census.mcm)[1:8],'.mcm')
```

Combining data
```{r}
msm.data.county <- full_join(county.data, county.srm, by = c('county1','tyear', 'c')) %>% full_join(county.mcm, by = c('county1','tyear', 'c'))

msm.data.census <- full_join(ct.data, census.srm, by = c('GEOID','tyear', 'c')) %>% full_join(census.mcm, by = c('GEOID','tyear','c'))
```

Adding states
```{r}
format.states <- function(data) {
  data %>% group_by(c) %>% 
  mutate(state.srm = case_when(tweek < epi.start.srm ~ 1,
                           tweek >= epi.start.srm & tweek < epi.end.srm ~ 2,
                           tweek >= epi.end.srm ~ 1,
                           is.na(epi.start.srm) ~ 1),
         state.mcm = case_when(tweek < epi.start.mcm ~ 1,
                           tweek >= epi.start.mcm & tweek < epi.end.mcm ~ 2,
                           tweek >= epi.end.mcm ~ 1,
                           is.na(epi.start.mcm) ~ 1))
}


msm.data.county <- format.states(msm.data.county)
msm.data.census <- format.states(msm.data.census)
```


Merging with climate data
```{r}
clim.county <- fread('Data/climate_county_2000_2023.csv')[,-1]
msm.data.county <- merge(clim.county, msm.data.county, by = c('county1','wkyr'))

clim.ct <- fread('Data/climate_2000_2023.csv', colClasses = c('GEOID' = 'character'))[,-1]
msm.data.census <- merge(clim.ct, msm.data.census, by = c('GEOID','wkyr'))
```


Writing
```{r}
msm.data.county %>% group_by(county1) %>%
  mutate(aseasonal = ifelse(all(is.na(epi.start.srm)),1,0)) %>%
  dplyr::filter(aseasonal == 1) %>% ungroup() %>% dplyr::select(county1, aseasonal) %>% distinct() %>% nrow()
msm.data.census %>% group_by(GEOID) %>%
  mutate(aseasonal = ifelse(all(is.na(epi.start.srm)),1,0)) %>%
  dplyr::filter(aseasonal == 1) %>% ungroup() %>% dplyr::select(GEOID, aseasonal) %>% distinct() %>% nrow()

write.csv(msm.data.county, 'Data/msm.data.county2023.csv')
write.csv(msm.data.census, 'Data/msm.data.census2023.csv')
```
